+++ packages/opencensus-core/test/test-root-span.ts
@@ -271,9 +271,7 @@ describe('RootSpan', () => {
 
       const rootSpan = new RootSpan(tracer, name, kind, traceId, parentSpanId);
       rootSpan.start();
-
-      rootSpan.addMessageEvent(
-          types.MessageEventType.UNSPECIFIED, 'message_event_test_id');
+      rootSpan.addMessageEvent(types.MessageEventType.UNSPECIFIED, 1);
 
       assert.ok(rootSpan.messageEvents.length > 0);
       assert.equal(rootSpan.droppedMessageEventsCount, 0);