+++ packages/opencensus-exporter-stackdriver/test/test-stackdriver-cloudtrace-utils.ts
@@ -98,17 +98,14 @@ describe('Stackdriver CloudTrace Exporter Utils', () => {
     ];
     const messageEvents: coreTypes.MessageEvent[] = [
       {
-        id: 'aaaa',
+        id: 1,
         timestamp: ts,
         type: coreTypes.MessageEventType.SENT,
         compressedSize: 100,
         uncompressedSize: 12
       },
-      {id: 'ffff', timestamp: ts, type: coreTypes.MessageEventType.RECEIVED}, {
-        id: 'eeee',
-        timestamp: ts,
-        type: coreTypes.MessageEventType.UNSPECIFIED
-      }
+      {id: 1, timestamp: ts, type: coreTypes.MessageEventType.RECEIVED},
+      {id: 1, timestamp: ts, type: coreTypes.MessageEventType.UNSPECIFIED}
     ];
 
     const expectedTimeEvent = [
@@ -147,22 +144,18 @@ describe('Stackdriver CloudTrace Exporter Utils', () => {
         }
       },
       {
-        messageEvent: {
-          compressedSize: '100',
-          id: 'aaaa',
-          type: 1,
-          uncompressedSize: '12'
-        },
+        messageEvent:
+            {compressedSize: '100', id: '1', type: 1, uncompressedSize: '12'},
         time: '1970-01-02T10:17:36.789Z',
       },
       {
         messageEvent:
-            {compressedSize: '0', id: 'ffff', type: 2, uncompressedSize: '0'},
+            {compressedSize: '0', id: '1', type: 2, uncompressedSize: '0'},
         time: '1970-01-02T10:17:36.789Z',
       },
       {
         messageEvent:
-            {compressedSize: '0', id: 'eeee', type: 0, uncompressedSize: '0'},
+            {compressedSize: '0', id: '1', type: 0, uncompressedSize: '0'},
         time: '1970-01-02T10:17:36.789Z',
       }
     ];