+++ packages/opencensus-instrumentation-http/src/http.ts
@@ -19,7 +19,6 @@ import {ClientRequest, ClientResponse, IncomingMessage, request, RequestOptions,
 import * as semver from 'semver';
 import * as shimmer from 'shimmer';
 import * as url from 'url';
-import * as uuid from 'uuid';
 import * as stats from './http-stats';
 import {IgnoreMatcher} from './types';
 
@@ -242,10 +241,7 @@ export class HttpPlugin extends BasePlugin {
 
             rootSpan.setStatus(
                 HttpPlugin.parseResponseStatus(response.statusCode));
-
-            // Message Event ID is not defined
-            rootSpan.addMessageEvent(
-                MessageEventType.RECEIVED, uuid.v4().split('-').join(''));
+            rootSpan.addMessageEvent(MessageEventType.RECEIVED, 1);
 
             tags.set(
                 stats.HTTP_SERVER_METHOD, {value: method},
@@ -407,10 +403,7 @@ export class HttpPlugin extends BasePlugin {
                 stats.HTTP_CLIENT_STATUS,
                 {value: response.statusCode.toString()});
           }
-
-          // Message Event ID is not defined
-          span.addMessageEvent(
-              MessageEventType.SENT, uuid.v4().split('-').join(''));
+          span.addMessageEvent(MessageEventType.SENT, 1);
 
           HttpPlugin.recordStats(span.kind, tags, Date.now() - startTime);
           span.end();