+++ packages/opencensus-instrumentation-http/test/test-http.ts
@@ -14,11 +14,12 @@
  * limitations under the License.
  */
 
-import {CoreTracer, globalStats, HeaderGetter, HeaderSetter, logger, Measurement, Propagation, RootSpan, Span, SpanContext, SpanEventListener, StatsEventListener, TagKey, TagMap, TagValue, View} from '@opencensus/core';
+import {CoreTracer, globalStats, HeaderGetter, HeaderSetter, logger, Measurement, MessageEventType, Propagation, RootSpan, Span, SpanContext, SpanEventListener, StatsEventListener, TagKey, TagMap, TagValue, View} from '@opencensus/core';
 import * as assert from 'assert';
 import * as http from 'http';
 import * as nock from 'nock';
 import * as shimmer from 'shimmer';
+
 import {HttpPlugin, plugin} from '../src/';
 import * as stats from '../src/http-stats';
 
@@ -223,6 +224,9 @@ describe('HttpPlugin', () => {
 
         const span = rootSpanVerifier.endedRootSpans[0];
         assertSpanAttributes(span, 200, 'GET', hostName, testPath);
+        assert.equal(span.messageEvents.length, 1);
+        assert.equal(span.messageEvents[0].type, MessageEventType.SENT);
+        assert.equal(span.messageEvents[0].id, '1');
         assertClientStats(testExporter, 200, 'GET');
       });
     });
@@ -265,6 +269,9 @@ describe('HttpPlugin', () => {
           assert.strictEqual(root.traceId, root.spans[0].traceId);
           const span = root.spans[0];
           assertSpanAttributes(span, 200, 'GET', hostName, testPath);
+          assert.equal(span.messageEvents.length, 1);
+          assert.equal(span.messageEvents[0].type, MessageEventType.SENT);
+          assert.equal(span.messageEvents[0].id, '1');
           assertClientStats(testExporter, 200, 'GET');
         });
       });
@@ -307,6 +314,9 @@ describe('HttpPlugin', () => {
             assert.strictEqual(root.spans.length, i + 1);
             assert.ok(root.spans[i].name.indexOf(testPath) >= 0);
             assert.strictEqual(root.traceId, root.spans[i].traceId);
+            assert.equal(root.spans[i].messageEvents[0].id, 1);
+            assert.equal(
+                root.spans[i].messageEvents[0].type, MessageEventType.SENT);
           });
         }
         assert.strictEqual(rootSpanVerifier.endedRootSpans.length, 0);